+++ src/BoclipsClient.ts
@@ -1,5 +1,6 @@
 import { CollectionsClient } from './sub-clients/collections/client/CollectionsClient';
 import { ContentPartnersClient } from './sub-clients/contentPartners/client/ContentPartnersClient';
+import { EventsClient } from './sub-clients/events/client/EventsClient';
 import { HttpFeedsClient } from './sub-clients/httpFeeds/client/HttpFeedsClient';
 import { LegalRestrictionsClient } from './sub-clients/legalRestrictions/client/LegalRestrictionsClient';
 import { OrdersClient } from './sub-clients/orders/client/OrdersClient';
@@ -12,4 +13,5 @@ export interface BoclipsClient {
   feedsClient: HttpFeedsClient;
   collectionsClient: CollectionsClient;
   ordersClient: OrdersClient;
+  eventsClient: EventsClient;
 }