+++ src/sub-clients/events/client/EventsClient.ts
@@ -1,0 +1,5 @@
+import { PageRenderedRequest } from '../model/PageRenderedRequest';
+
+export interface EventsClient {
+  trackPageRendered(request: PageRenderedRequest): Promise<{}>;
+}