+++ src/sub-clients/events/model/PageRenderedRequest.ts
@@ -1,0 +1,5 @@
+import { EventRequest } from './EventRequest';
+
+export interface PageRenderedRequest extends EventRequest {
+  url: string;
+}